Conflicts:
	src/Engine/Sprite.cpp

- Fixed most of Visual Studio compiler warnings
- Also checked project trough cppcheck an fixed warnings

+++ include/UtH/Engine/DefaultScene.hpp
@@ -19,7 +19,7 @@ namespace uth
 		virtual bool Init();
 		virtual bool DeInit();
 
-		virtual bool Update(double dt);
+		virtual bool Update(float dt);
 		virtual bool Draw();
 	protected:
 		Shader* shader;
